backend/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "midi-ai-api"
version = "1.0.0"
description = "FastAPI backend for AI melody generation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "torch>=2.5.0",
    "music21>=9.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
    "httpx>=0.27.0",
    "midi2audio>=0.1.1",
    "miditok>=3.0",
    "symusic>=0.5.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "pytest-timeout>=2.3.0",
]
dev = [
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
